+++ test/database/uuid_primary_key_test.rb
@@ -2,10 +2,10 @@
 
 require "test_helper"
 
-class UuidPrimaryKeyTest < RailsFusion::TestCase
+class UuidPrimaryKeyTest < HybridSearch::TestCase
   def test_search_works_with_uuid_primary_key
     doc = FusionDoc.create!(name: "UUID doc", description: "identifies by uuid")
-    provider = FusionDoc.rails_fusion_definition.embedding_config.provider
+    provider = FusionDoc.hybrid_search_definition.embedding_config.provider
     doc.update_columns(search_embedding: provider.embed(["UUID doc identifies by uuid"]).first)
 
     results = FusionDoc.fusion_search("UUID doc")