+++ test/ranking/recency_test.rb
@@ -4,7 +4,7 @@ require "test_helper"
 
 class RecencyTest < Minitest::Test
   def test_nil_timestamp_gets_no_boost
-    assert_equal 0.0, RailsFusion::Ranking::Recency.contribution(nil, half_life: 86_400, rrf_k: 60)
+    assert_equal 0.0, HybridSearch::Ranking::Recency.contribution(nil, half_life: 86_400, rrf_k: 60)
   end
 
   def test_newer_result_gets_larger_bounded_boost
@@ -12,8 +12,8 @@ class RecencyTest < Minitest::Test
     newer = clock - 60
     older = clock - (10 * 86_400)
 
-    newer_score = RailsFusion::Ranking::Recency.contribution(newer, half_life: 86_400, rrf_k: 60, clock: clock)
-    older_score = RailsFusion::Ranking::Recency.contribution(older, half_life: 86_400, rrf_k: 60, clock: clock)
+    newer_score = HybridSearch::Ranking::Recency.contribution(newer, half_life: 86_400, rrf_k: 60, clock: clock)
+    older_score = HybridSearch::Ranking::Recency.contribution(older, half_life: 86_400, rrf_k: 60, clock: clock)
 
     assert_operator newer_score, :>, older_score
     assert_operator newer_score, :<=, 1.0 / 60
@@ -22,8 +22,8 @@ class RecencyTest < Minitest::Test
   def test_deterministic_with_frozen_clock
     clock = Time.at(2_000_000)
     ts = clock - 3600
-    a = RailsFusion::Ranking::Recency.contribution(ts, half_life: 86_400, rrf_k: 60, clock: clock)
-    b = RailsFusion::Ranking::Recency.contribution(ts, half_life: 86_400, rrf_k: 60, clock: clock)
+    a = HybridSearch::Ranking::Recency.contribution(ts, half_life: 86_400, rrf_k: 60, clock: clock)
+    b = HybridSearch::Ranking::Recency.contribution(ts, half_life: 86_400, rrf_k: 60, clock: clock)
     assert_equal a, b
   end
 end