+++ test/test_helper.rb
@@ -4,7 +4,7 @@ $LOAD_PATH.unshift File.expand_path("../lib", __dir__)
 
 require "active_record"
 require "active_job"
-require "rails_fusion"
+require "hybrid_search"
 require "minitest/autorun"
 
 ActiveJob::Base.queue_adapter = :test
@@ -15,20 +15,20 @@ ActiveRecord::Base.establish_connection(
   host: ENV.fetch("PGHOST", "localhost"),
   port: ENV.fetch("PGPORT", "5433").to_i,
   username: ENV.fetch("PGUSER", ENV.fetch("USER", nil)),
-  database: ENV.fetch("PGDATABASE", "rails_fusion_test")
+  database: ENV.fetch("PGDATABASE", "hybrid_search_test")
 )
 ActiveRecord::Base.logger = nil
-ActiveRecord::Base.include(RailsFusion::Model) unless ActiveRecord::Base < RailsFusion::Model
+ActiveRecord::Base.include(HybridSearch::Model) unless ActiveRecord::Base < HybridSearch::Model
 
 require_relative "support/schema"
 require_relative "support/models"
 
-module RailsFusion
+module HybridSearch
   class TestCase < Minitest::Test
     TABLES = %w[fusion_products fusion_articles fusion_docs].freeze
 
     def setup
-      RailsFusion.reset_configuration!
+      HybridSearch.reset_configuration!
       ActiveJob::Base.queue_adapter.enqueued_jobs.clear
       ActiveJob::Base.queue_adapter.performed_jobs.clear
     end